Slight code re-org, usage comment and better naming of test file.
